+++ imageutils/tensor.go
@@ -35,7 +35,7 @@ func ToTensorFromDirectory(dirPath string, loadOpts LoadOpts, tensorOpts ToTenso
 		return nil, fmt.Errorf("TargetSize must be defined")
 	}
 
-	backing := []float64{}
+	backing := []float32{}
 	imagesCount := 0
 
 	err := filepath.WalkDir(dirPath, func(path string, d fs.DirEntry, err error) error {
@@ -63,7 +63,7 @@ func ToTensorFromDirectory(dirPath string, loadOpts LoadOpts, tensorOpts ToTenso
 	}
 
 	return tensor.New(
-		tensor.Of(tensor.Float64),
+		tensor.Of(tensor.Float32),
 		tensor.WithShape(imagesCount, 3, int(loadOpts.TargetSize[0]), int(loadOpts.TargetSize[1])), // count, channels, width, height
 		tensor.WithBacking(backing),
 	), nil
@@ -74,18 +74,18 @@ func ToTensor(img image.Image, opts ToTensorOpts) tensor.Tensor {
 	bounds := img.Bounds()
 
 	return tensor.New(
-		tensor.Of(tensor.Float64),
+		tensor.Of(tensor.Float32),
 		tensor.WithShape(1, 3, bounds.Max.X, bounds.Max.Y), // batchSize, channels, width, height
 		tensor.WithBacking(ToArray(img, opts)),
 	)
 }
 
-// ToArray converts the image in a []float64
-func ToArray(img image.Image, opts ToTensorOpts) []float64 {
+// ToArray converts the image in a []float32
+func ToArray(img image.Image, opts ToTensorOpts) []float32 {
 	bounds := img.Bounds()
 	width, height := bounds.Max.X, bounds.Max.Y
 
-	pixels := make([]float64, 3*width*height)
+	pixels := make([]float32, 3*width*height)
 
 	for x := 0; x < width; x++ {
 		for y := 0; y < height; y++ {
@@ -100,31 +100,31 @@ func ToArray(img image.Image, opts ToTensorOpts) []float64 {
 	return pixels
 }
 
-func pixelWeight(pixel color.Color, opts ToTensorOpts) (float64, float64, float64) {
+func pixelWeight(pixel color.Color, opts ToTensorOpts) (float32, float32, float32) {
 	r, g, b, _ := pixel.RGBA()
 
 	switch opts.TensorMode {
 	case TensorModeTensorFlow:
 		// https://github.com/tensorflow/tensorflow/blob/v2.4.1/tensorflow/python/keras/applications/imagenet_utils.py#L192
 
-		return float64(r/256)/127.5 - 1.0,
-			float64(g/256)/127.5 - 1.0,
-			float64(b/256)/127.5 - 1.0
+		return float32(r/256)/127.5 - 1.0,
+			float32(g/256)/127.5 - 1.0,
+			float32(b/256)/127.5 - 1.0
 	case TensorModeTorch:
 		// https://github.com/tensorflow/tensorflow/blob/v2.4.1/tensorflow/python/keras/applications/imagenet_utils.py#L197
-		mean := []float64{0.485, 0.456, 0.406}
-		std := []float64{0.229, 0.224, 0.225}
+		mean := []float32{0.485, 0.456, 0.406}
+		std := []float32{0.229, 0.224, 0.225}
 
-		return (float64(r)/65536 - mean[0]) / std[0],
-			(float64(g)/65536 - mean[1]) / std[1],
-			(float64(b)/65536 - mean[2]) / std[2]
+		return (float32(r)/65536 - mean[0]) / std[0],
+			(float32(g)/65536 - mean[1]) / std[1],
+			(float32(b)/65536 - mean[2]) / std[2]
 	default:
 		// https://github.com/tensorflow/tensorflow/blob/master/tensorflow/python/keras/applications/imagenet_utils.py#L202
-		mean := []float64{103.939, 116.779, 123.68}
+		mean := []float32{103.939, 116.779, 123.68}
 
 		// RGB -> BGR
-		return float64(b/256) - mean[0],
-			float64(g/256) - mean[1],
-			float64(r/256) - mean[2]
+		return float32(b/256) - mean[0],
+			float32(g/256) - mean[1],
+			float32(r/256) - mean[2]
 	}
 }
