+++ tabnet/classifier.go
@@ -77,7 +77,7 @@ func (r *Classifier) Model() *godl.Model {
 
 func (r *Classifier) Train(trainX, trainY, validateX, validateY tensor.Tensor, opts godl.TrainOpts) error {
 	if opts.CostFn == nil {
-		lambdaSparse := gorgonia.NewConstant(float64(1e-3))
+		lambdaSparse := gorgonia.NewConstant(float32(1e-3))
 		opts.CostFn = func(output *gorgonia.Node, innerLoss *gorgonia.Node, y *gorgonia.Node) *gorgonia.Node {
 			cost := godl.CategoricalCrossEntropyLoss(output, y, godl.CrossEntropyLossOpt{})
 			cost = gorgonia.Must(gorgonia.Sub(cost, gorgonia.Must(gorgonia.Mul(lambdaSparse, innerLoss))))
